Implements automake-like testing; bugfixes
Automake-like testing is mainly to enable feature detection. This comes
hard into play with, notably, <filesystem>. <filesystem> is a really,
really dumb header that in some versions requires another library to
be included. This mainly applies no clang on non-apple devices, and gcc.
This is also bound to the version of the stdlib, which adds for more of
a challenge, and this can't just be done with a linear test. MSVC
doesn't require linking at all.

Some code compilation tests checks for compatibility makes it possible
to determine exactly what is needed, mostly. This excludes
<experiemental/filesystem> in the built-in method, but this can easily
be extended further with custom methods and checks, whether implemented
by the library, SCons, or you.

Docs coming in 6-8:tm: